Specify Encoder for APIDriver
Remove the JsonApiDriver class and change the APIDriver class to allow a
custom input encoder and response body decoder as well as default
request headers to be be set on the subclass of APIDriver in the test
implementation.

This simplifies Kookaburra's code base somewhat by removing an
unnecessary layer of abstraction, and it makes it easier for users of
Kookaburra to develop against non-JSON APIs without having to either
wait for Kookaburra to support them or understand enough about the way
the APIDriver works to wrap it in the same way that JsonApiDriver did.