Version bump to 1.0.0
The JsonApiDriver class was removed in favor of just allowing customer
data encoder/decoder routines and headers to be set on subclasses of
Kookaburra::APIDriver. Existing subclasses of JsonApiDriver should be
easy to update by changing a class that currently looks like:

  require 'kookaburra/json_api_driver'
  class MyAPIDriver < Kookaburra::JsonApiDriver
    # all my api methods
    # ...
  end

to look like:

  require 'json'
  require 'kookaburra/api_driver'
  class MyAPIDriver < Kookaburra::APIDriver
    encode_with { |data| JSON.dump(data) }
    decode_with { |data| JSON.parse(data) }
    header 'Content-Type', 'application/json'
    header 'Accept', 'application/json'

    # all my api methods
    # ...
  end

Note also that the optional `:expected_response_status` parameter to the
APIDriver's get, post, put and delete methods is no longer supported.
Instead, and UnexpectedResponse exception will now be raised if the
response status in not in the 2xx or 3xx range. Additionally, the
APIDriver will now follow redirects.