Removes Kookaburra::MentalModelMatcher
The matcher promoted a style of testing where one assumes that the mental model
contains everything that the application knows about. If the application, say,
has widgets A, B and C, but a particular test only creates widgets A and B, then
the matcher would fail if you asked it to match agains all the widgets returned
by the application. This would be fine in the traditional Rails testing approach
of using a testing data base and cleaning all data between tests, but
Kookaburra's design goal is to allow it to run tests against an application  over
which it has no direct data-store control.