Disprove(?) the anonymous-module-based solution...
I thought I had a solution to the test that was failing in WIP, but no!
As it turned out, the tests were passing or failing when run in
different orders.  After puzzling through for a while, the facepalm  hit
me:  *creating* dynamic modules is perfectly fine.  You just can't ever
get rid of them again, because Ruby has no facility for unincluding a
module.  They'll just hang around forever.  We could address this by
undefining methods again, but that's extra work that we'll have to
remember to do again later.

With that option off the table, the only other solution I can think of
is method_missing on TestHelpers that delegates to
Kookaburra.configuration.  That is, of course, if we want to support the
top-level syntax in the first place.  Another alternative would be
something like:

  apps[:alice].api.create_widget # ...
  apps[:bob].api.create_widget # ...